# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TrossApp API",
    "version": "1.0.0",
    "description": "\n        TrossApp Backend REST API - Skills-based work order management system.\n        \n        This documentation covers foundational authentication, user management, \n        and role-based access control endpoints.\n        \n        ## Authentication\n        \n        Most endpoints require a Bearer token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Dual Authentication Strategy\n        \n        - **Development Mode:** JWT-based authentication with test users\n        - **Production Mode:** Auth0 OAuth2/OIDC integration\n        \n        ## Response Format\n        \n        All successful responses follow this format:\n        ```json\n        {\n          \"success\": true,\n          \"data\": {...},\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n        \n        Error responses:\n        ```json\n        {\n          \"success\": false,\n          \"error\": \"Error message describing what went wrong\",\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n      ",
    "contact": {
      "name": "TrossApp Team",
      "email": "dev@trossapp.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
...
    },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/dev/token (dev) or Auth0 (prod)"
      }
    },
...
  },
...
  "paths": {
...
    "/api/schema": {
      "get": {
        "tags": [
          "Schema"
        ],
        "summary": "Get all available tables",
        "description": "Returns list of tables in public schema with metadata",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of tables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schema/{tableName}": {
      "get": {
        "tags": [
          "Schema"
        ],
        "summary": "Get table schema metadata",
        "description": "Returns complete schema info for auto-generating UI (columns, types, constraints, relationships)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableName",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Table name (e.g., 'users', 'roles')"
          }
        ],
        "responses": {
          "200": {
            "description": "Table schema metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tableName": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "nullable": {
                                "type": "boolean"
                              },
                              "uiType": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "readonly": {
                                "type": "boolean"
                              },
                              "foreignKey": {
                                "type": "object"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Table not found"
          }
        }
      }
    },
    "/api/schema/{tableName}/options/{column}": {
      "get": {
        "tags": [
          "Schema"
        ],
        "summary": "Get select options for foreign key field",
        "description": "Returns {value, label} pairs for dropdown/select fields",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tableName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "column",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Column name (e.g., 'role_id')"
          }
        ],
        "responses": {
          "200": {
            "description": "Select options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "integer"
                          },
                          "label": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}